optimisation: when fetching a ticket, it will now
1. create the html based on data from database
2. fetch data on the remote for that ticket only
3. create the html based on data received at two
4. send 3 if it differs to 1
5. terminate request.

Before, the step 2 was done by synchronising all the tickets which were
out of data, and then recreating the html based on data from the newly
synchronised data. Unfortunately, as long as we can't synchronise a
single ticket without interfering with the algorithm that detects the
tickets which are out of date, the newer step 2, is significantly faster.
A fetch-ticket request now takes about 1s to finish, vs 7s before.