Retrieve up to date attachment list using a jira REST request to issue.
when running a request to get the list of attachment belonging to a ticket,
the code used to synchronise the database to ensure it could return the
up-to-date data. This meant the request would take relatively long to
finish. Instead, it now relies on retrieving only the data for one ticket
and only synchronise the database if that isn't enough. In other words, in
the majority of cases, the fetch requests finishes in about one second,
instead of seven. And in a few cases, it takes one second longer.